Struct rotor_carbon::Sink

pub struct Sink<C, S>(_, _) where S: ActiveStream;

This is a wrapper around the machinery to send data

Use sink.sender() go to get an actual object you may send to

Note sink.sender() holds lock on the underlying buffer and doesn't send data, until sender is dropped. This is useful for sending data in single bulk.

Methods

impl<C, S: ActiveStream> Sink<C, S> where S::Address: Debug

fn sender(&self) -> Sender<C, S>

Get the Sender object to send data

Rules of thumb:

Hold it only for sending data, not for fetching actual metrics

If you poll for metrics in the loop, drop the sender while doing sleep()